fixed formatting, rename main entry, changed include some type hinting,
